requirements.txt
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
kiwisolver==1.4.7
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
scipy==1.14.1
six==1.17.0
